cleanups, tests
