+++ logstage/logstage-core/src/main/scala/izumi/logstage/api/IzLogger.scala
@@ -2,7 +2,7 @@ package izumi.logstage.api
 
 import izumi.logstage.api.Log.CustomContext
 import izumi.logstage.api.logger.{AbstractMacroLogger, LogRouter, RoutingLogger}
-import izumi.logstage.api.rendering.LogstageCodec
+import izumi.logstage.api.rendering.AnyEncoded
 
 class IzLogger
 (
@@ -27,36 +27,6 @@ object IzLogger extends IzLoggerConvenienceApi {
   }
 }
 
-sealed trait AnyEncoded {
-  type T
-  def value: T
-  def codec: Option[LogstageCodec[T]]
-}
-sealed trait StrictEncoded extends AnyEncoded
 
-trait StrictEncodedLow {
-  @inline implicit def to[A](a: A)(implicit maybeCodec: LogstageCodec[A]): StrictEncoded = new StrictEncoded {
-    override type T = A
 
-    override def value: A = a
 
-    override def codec: Option[LogstageCodec[A]] = Option(maybeCodec)
-  }
-
-}
-
-object StrictEncoded extends StrictEncodedLow {
-
-  @inline implicit def toPair[A](a: (String, A))(implicit maybeCodec: LogstageCodec[A]): (String, StrictEncoded) = (a._1, to(a._2))
-}
-
-trait AnyEncodedLow {
-  @inline implicit def to[A](a: A)(implicit maybeCodec: LogstageCodec[A] = null): AnyEncoded = StrictEncoded.to(a)
-
-}
-
-object AnyEncoded extends AnyEncodedLow {
-
-  @inline implicit def toPair[A](a: (String, A))(implicit maybeCodec: LogstageCodec[A] = null): (String, AnyEncoded) = (a._1, to(a._2))
-
-}
